LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/specifier-error-expected.txt

CONSOLE MESSAGE: TypeError: Module specifier, 'string-without-dot-slash-prefix' does not start with "/", "./", or "../". Referenced from http://web-platform.test:8800/html/semantics/scripting-1/the-script-element/module/bad-module-specifier.js

PASS Test that invalid module specifier leads to TypeError on window.
